Changes due to pandas update (#660)
* draft

* draft

* draft

* draft

* draft

* draft

* draft

* final rework pandas warnings

* fix: remove changes for older pandas versions

* fix: uncomment internal datasets tests

* chore: update changelog

* fix: convert target to float32

* fix: convert target to float64

* fix: fix tests

* chore: lint

* fix: update internal dataset hash

* add: update docstrings of transforms and methods

* add: process wide df with different target dtypes

* fix: update docstrings

* fix: fix test

* fix: update docstrings

* fix: update docstrings

* fix: fix test

---------

+++ tests/test_transforms/test_math/test_exog_shift_transform.py
@@ -183,9 +183,9 @@ def test_transformed_names(ts_name, lag, horizon, expected, request):
 @pytest.mark.parametrize(
     "lag,horizon,expected_types",
     (
-        (1, None, {"feat1_shift_1": "float", "feat2_shift_1": "float", "feat3_shift_1": "float", "target": "int"}),
-        ("auto", 1, {"feat1_shift_1": "float", "feat2_shift_2": "float", "feat3": "int", "target": "int"}),
-        ("auto", 2, {"feat1_shift_2": "float", "feat2_shift_3": "float", "feat3_shift_1": "float", "target": "int"}),
+        (1, None, {"feat1_shift_1": "float", "feat2_shift_1": "float", "feat3_shift_1": "float", "target": "float"}),
+        ("auto", 1, {"feat1_shift_1": "float", "feat2_shift_2": "float", "feat3": "int", "target": "float"}),
+        ("auto", 2, {"feat1_shift_2": "float", "feat2_shift_3": "float", "feat3_shift_1": "float", "target": "float"}),
     ),
 )
 @pytest.mark.parametrize(
